Store: If TaxJar plugin is active, do not show core settings
